refactor: enhance type safety by introducing handler protocols and cleaning up imports
